pgs (field presence): split proto3 tests
This change splits the testing harness for proto3
field presence, so we can retain functional test
coverage for older protoc versions that do not
support proto3 field presence. IE >v3.15.0
